Add text + round avatar's corners
Also change the printout size to 2480x3508 hard-coded, which corresponds
A4 at 300dpi. This hack solves font sizes problem: if the image size is
not fixed, the font size will differ from image to image and the only
solution to get a text label of the same size is to brute-force font
sizes, which is not cool. So this hack is better. Also, it guarantees
the correctness of an output size since it does not depend on the
background size anymore.

Note: the printout result is served incorrectly right now, IDK why.